restructure bookings module
