+++ apps/api/v2/src/modules/bookings/booking.repository.ts
@@ -1,4 +1,4 @@
-import { CreateBookingInput } from "@/modules/booking/input/create-booking";
+import { CreateBookingInput } from "@/modules/bookings/inputs/create-booking";
 import { PrismaReadService } from "@/modules/prisma/prisma-read.service";
 import { PrismaWriteService } from "@/modules/prisma/prisma-write.service";
 import { Injectable } from "@nestjs/common";
