+++ apps/api/v2/src/modules/bookings/inputs/create-booking.ts
